+++ ProjetoDA/models/Orcamento.cs
@@ -10,17 +10,18 @@ namespace ProjetoDA
     public class Orcamento
     {
         public int Id { get; set; }
-        public int ValorMaximo { get; set; } // No diagrama está definido como 'int'
+        // Passa a decimal para suportar cêntimos
+        public decimal ValorMaximo { get; set; }
         public int Mes { get; set; }
         public int Ano { get; set; }
 
-        // Relações com Utilizador indicadas no diagrama (user cria, user altera)
         public Utilizador UserCria { get; set; }
         public Utilizador UserAltera { get; set; }
 
         public Orcamento() { }
 
-        public Orcamento(int valorMaximo, int mes, int ano, Utilizador userCria)
+        // Alterar também aqui no construtor
+        public Orcamento(decimal valorMaximo, int mes, int ano, Utilizador userCria)
         {
             ValorMaximo = valorMaximo;
             Mes = mes;
